+++ src/pages/index.ts
@@ -1,2 +1,2 @@
-export { default as CardEditorPage } from "./CardEditorPage";
-export { default as CardScenePage } from "./CardScenePage";
+export { default as CardEditorPage } from "./CardEditorPage/CardEditorPage";
+export { default as CardScenePage } from "./CardScenePage/CardScenePage";
